requirements.txt
discord==2.0.0
googletran==3.1.0a0